+++ src/DefinitionValue.php
@@ -1,0 +1,21 @@
+<?php declare(strict_types=1);
+
+namespace HbLib\Container;
+
+class DefinitionValue extends AbstractDefinition
+{
+    /**
+     * @var mixed
+     */
+    private $value;
+
+    public function __construct($value)
+    {
+        $this->value = $value;
+    }
+
+    public function getValue()
+    {
+        return $this->value;
+    }
+}